# src/openapi.yaml
openapi: 3.0.3
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform built on the Stellar
    blockchain (Soroban smart contracts).
  version: 1.0.0
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: API v1
  - url: /api/v2
    description: API v2 (includes SSE stream endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform
        operations covered by their granted scopes (#1019). Legacy keys without a scope list are
        unrestricted.
...
paths:
...
  /scouts/{wallet}/recommendations:
    get:
      tags:
        - scouts
      summary: Weighted, cursor-paginated player recommendations for this scout.
      operationId: getScoutRecommendations
      description: |-
        Weighted player recommendations derived from the scout's saved searches,
        bookmarks, and past contact unlocks (falls back to top-tier players when
        the scout has no history). Cursor-paginated; results cached 10 minutes.
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
          description: Scout's Stellar public key
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Optional — last player_id of the previous page
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
          description: Optional — default 20
      responses:
        '200':
          description: >-
            { success: true, data: PlayerSummary[], nextCursor, meta: { preferredRegion,
            preferredPosition, minTier, totalCandidates } }
        '400':
          description: '{ success: false, error: string } - Invalid query parameters'
      security:
        - bearerAuth: []
        - apiKeyAuth:
            - read:recommendations
      x-required-role:
        - scout
      x-api-key-scope: read:recommendations
      x-owner-only: true
...